@@ -2,13 +2,13 @@ package com.bank.podo.domain.account.entity;
 
 import com.bank.podo.domain.account.enums.AccountType;
 import com.bank.podo.domain.user.entity.User;
+import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 import org.hibernate.annotations.ColumnDefault;
 
-import javax.persistence.*;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
