@@ -1,11 +1,11 @@
 package com.bank.podo.domain.user.entity;
 
 import com.bank.podo.domain.user.enums.Role;
+import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
