@@ -1,5 +1,5 @@
 package com.bank.podo.domain.user.enums;
 
 public enum Role {
-    ROLE_DELETED, ROLE_USER, ROLE_ADMIN
+    ROLE_DELETED, ROLE_USER, ROLE_MANAGER, ROLE_ADMIN
 }
